+++ .eslintrc.cjs
@@ -1,7 +1,7 @@
 module.exports = {
     env: {
         browser: true,
-        es2021: true,
+        es2022: true,
     },
     parser: '@typescript-eslint/parser',
     parserOptions: {