+++ packages/app/src/views/Realm/data.ts
@@ -1,4 +1,4 @@
-import { MoveStrategy, NoBloodSwitchLink, Operator, SkillNameMatch, generateStrategy } from 'sa-core';
+import { MoveStrategy, NoBloodSwitchLink, Operator, SkillNameMatch, generateStrategyByName } from 'sa-core';
 import * as SABattle from 'sa-core/battle';
 import * as SAEntity from 'sa-core/entity';
 
@@ -12,7 +12,7 @@ interface LevelPetsData {
 const data: LevelPetsData = {
     LevelXTeamRoom: {
         cts: [1656056275, 1656945596, 1657863632],
-        strategy: generateStrategy(['幻梦芳逝', '剑挥四方', '破寂同灾'], ['蒂朵', '六界帝神', '深渊狱神·哈迪斯']),
+        strategy: generateStrategyByName(['幻梦芳逝', '剑挥四方', '破寂同灾'], ['蒂朵', '六界帝神', '深渊狱神·哈迪斯']),
     },
     LevelExpTraining: {
         cts: [1656092908, 1656055512, 1656056275],
@@ -43,7 +43,7 @@ const data: LevelPetsData = {
     },
     LevelStudyTraining: {
         cts: [1656055512, 1655101462, 1656302059, 1657039061],
-        strategy: generateStrategy(['有女初长成', '王·龙子盛威决', '竭血残蝶'], ['幻影蝶', '艾欧丽娅', '千裳']),
+        strategy: generateStrategyByName(['有女初长成', '王·龙子盛威决', '竭血残蝶'], ['幻影蝶', '艾欧丽娅', '千裳']),
     },
     LevelCourageTower: {
         cts: [1656055512, 1656302059, 1656056275, 1656092908],
@@ -65,14 +65,14 @@ const data: LevelPetsData = {
     },
     LevelTitanHole: {
         cts: [1656055512, 1656945596, 1657863632, 1655101462, 1656092908],
-        strategy: generateStrategy(
+        strategy: generateStrategyByName(
             ['竭血残蝶', '剑挥四方', '破寂同灾', '神灵救世光', '有女初长成'],
             ['幻影蝶', '六界帝神', '深渊狱神·哈迪斯']
         ),
     },
     LevelElfKingsTrial: {
         cts: [1656055512, 1656945596, 1657863632],
-        strategy: generateStrategy(['竭血残蝶', '剑挥四方', '破寂同灾'], ['幻影蝶', '六界帝神', '深渊狱神·哈迪斯']),
+        strategy: generateStrategyByName(['竭血残蝶', '剑挥四方', '破寂同灾'], ['幻影蝶', '六界帝神', '深渊狱神·哈迪斯']),
     },
 };