+++ changelog.md
@@ -7,24 +7,28 @@ RoadMap: **见Readme**
 
 **todo已经迁移至issue/project**
 
+- [ ] 项目更名为SEA(Seerh5 Assistant Project), 启动器部分名称更改为SAEL(Seerh5 Assistant Launcher)
+
 # Core 当前版本 v0.5.0
 
 - [x] 移除chalk
 - [x] 将log逻辑全部移出到app
 - [x] 添加egret的license
 - [x] `LevelManager`关卡强制终止时现在会在当前对战结束后立即停止
-- [ ] 重构战斗策略模块概念
+- [x] 更换对战相关Hook名称, 向后微调`battleStart`(原`panelReady`)的节点
+- [x] 重构对战策略模块概念, 精简核心的职能, `resolveStrategy`由`Manager`导出, 主要逻辑的处理移交登录器层
+- [ ] 添加战斗单元测试
 
 - [ ] 涉及缩写sa的字符串部分用常量代替（缩写未来将要更改为sea）
-- [ ] 添加战斗单元测试
 - [ ] 整合官方的一个module显示log
   - [ ] 屏蔽官方实现
   - [ ] 更新hook位置
 - [ ] 属性攻击/特攻/物攻的枚举
+- [ ] 添加扩展支持
 
 # Core v0.6.0
 
-- [ ] 将type作为打包输出的一部分而是不全部包共享
+- [ ] 将type作为打包输出的一部分而是在monorepo内共享
 - [ ] 分离开发调试用的模块到专门的开发包
 
 # App 当前版本 v0.5.1